Add bug fixes so that a314disk is again romable
When switching version of vbcc some time in the past, the syntax for
assembly files were changed so that it is not allowed to have spaces
in expressions. This then introduced a bug which went unnoticed until
now. With these fixes it is again possible to add a314.device and
a314disk.device to the kickstart rom image, and that way boot from rom.